Use single settingName similar to MCM helper
INI files are case sensitive, but Papyrus does not always preserve the capitalization of string literals. This is possibly related to translation/known strings and is more common on shorter and common words. By combining section and key, the likelhood of capitalization changes is reduced. At the same time the same argument can be used for PapyrusINI and MCM helper, so also capitalization changes will be the same.